requirements.txt
hypercorn[trio]
fastapi[standard]
itsdangerous
lxml
python-jose[cryptography]
saxonche
xmlschema
requests
jinja2
python-multipart
pydantic_settings
filelock
ollama
anthropic
google-genai
mempalace
zxing-cpp
pymupdf
pytest
ruff
mypy
lxml-stubs
